Refactor
Added try-catch'es and ModelState validation

+++ Controllers/ItemListsController.cs
@@ -35,6 +35,8 @@ namespace ShappingList.Controllers
         public IActionResult Create([FromBody]ItemListModel model)
         {
 
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
 
             // map model to entity
             var itemList = _mapper.Map<ItemList>(model);
@@ -59,14 +61,24 @@ namespace ShappingList.Controllers
         [HttpGet]
         public IActionResult GetAll()
         {
-            var itemLists = _itemListService.GetAll();
-            return Ok(itemLists);
+            try
+            {
+                var itemLists = _itemListService.GetAll();
+                return Ok(itemLists);
+            }
+            catch(AppException ex) 
+            { 
+                return BadRequest(new { message = ex }); 
+            }
         }
 
         //! [Authorize]
         [HttpPut("{id}")]
         public IActionResult Update(int id, [FromBody]ItemListUpdateModel model)
         {
+
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
 
             // map model to entity and set id
             var itemList = _mapper.Map<ItemList>(model);
@@ -91,16 +103,34 @@ namespace ShappingList.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
-            _itemListService.Delete(id);
-            return Ok();
+            
+
+            try
+            {
+                _itemListService.Delete(id);
+                return Ok();
+            }
+            catch(AppException ex) 
+            { 
+                return BadRequest(new { message = ex }); 
+            }
         }
 
         //! [Authorize]
         [HttpGet("{id}")]
         public IActionResult GetById(int id)
         {
-            var itemList = _itemListService.GetById(id);
-            return Ok(itemList);
+            
+
+            try
+            {
+                var itemList = _itemListService.GetById(id);
+                return Ok(itemList);
+            }
+            catch(AppException ex) 
+            { 
+                return BadRequest(new { message = ex }); 
+            }
         }
 
 
@@ -108,31 +138,58 @@ namespace ShappingList.Controllers
         [HttpPost("{id}/items/new")]
         public IActionResult AddItem(int id, [FromBody]ItemModel model)
         {
+
+
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var item = _mapper.Map<Item>(model);
-            var list = _itemListService.GetById(id);
+
+            try
+            {
+                var list = _itemListService.GetById(id);
 
 
-            //? Maybe there's better way to assign list to item than on a controller side. 
-            //? Something better than override for Create method that takes list argument?
+                //? Maybe there's better way to assign list to item than on a controller side. 
+                //? Something better than override for Create method that takes list argument?
 
-            //TODO: Cover all endpoints, nulls etc. in controllers and services later on! 
-            //TODO: Try-catches! 
-            item.List = list;
-
-
-            _itemService.Create(item);
+                //TODO: Cover all endpoints, nulls etc. in controllers and services later on! 
+                //TODO: Try-catches! 
+                item.List = list;
 
 
-            return Ok(list);
+                _itemService.Create(item);
+
+
+                return Ok(list);
+            }
+            catch(AppException ex) 
+            { 
+                return BadRequest(new { message = ex }); 
+            }
+
+            
         }
 
         [HttpDelete]
         public IActionResult DeleteItem([FromQuery]int listId, [FromQuery]int itemId)
         {
+
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            try
+            {
+
             var list = _itemListService.GetById(listId);
             _itemService.Delete(itemId);
 
             return Ok(list);
+            }
+            catch(AppException ex)
+            {
+                return BadRequest(new { message = ex });
+            }
         }
     }
 }
